Rename OpenTabletDriverGUI to OpenTabletDriver
