+++ OpenTabletDriver.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 15.0.26124.0
 MinimumVisualStudioVersion = 15.0.26124.0
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TabletDriverLib", "TabletDriverLib\TabletDriverLib.csproj", "{E1407092-4EC3-4FDB-A1F7-27684B889750}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenTabletDriverGUI", "OpenTabletDriverGUI\OpenTabletDriverGUI.csproj", "{5720DD79-C5B1-410A-B059-AC385F702DA5}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenTabletDriver", "OpenTabletDriver\OpenTabletDriver.csproj", "{5720DD79-C5B1-410A-B059-AC385F702DA5}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NativeLib", "NativeLib\NativeLib.csproj", "{57A5101A-20F4-41AA-8649-2C3785DD5BA2}"
 EndProject
