+++ OpenTabletDriver/App.xaml
@@ -1,7 +1,7 @@
 <Application xmlns="https://github.com/avaloniaui"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:OpenTabletDriverGUI"
-             x:Class="OpenTabletDriverGUI.App">
+             xmlns:local="clr-namespace:OpenTabletDriver"
+             x:Class="OpenTabletDriver.App">
     <Application.DataTemplates>
         <local:ViewLocator/>
     </Application.DataTemplates>
