+++ OpenTabletDriver/Views/ConfigurationManager.xaml
@@ -1,10 +1,10 @@
 <Window xmlns="https://github.com/avaloniaui"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:vm="clr-namespace:OpenTabletDriverGUI.ViewModels;assembly=OpenTabletDriverGUI"
+        xmlns:vm="clr-namespace:OpenTabletDriver.ViewModels;assembly=OpenTabletDriver"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-        x:Class="OpenTabletDriverGUI.Views.ConfigurationManager"
+        x:Class="OpenTabletDriver.Views.ConfigurationManager"
         Width="1024" Height="768"
         Title="Configuration Manager">
     <Design.DataContext>
