+++ OpenTabletDriver/Views/ConfigurationManager.xaml.cs
@@ -1,7 +1,7 @@
 using Avalonia.Controls;
 using Avalonia.Markup.Xaml;
 
-namespace OpenTabletDriverGUI.Views
+namespace OpenTabletDriver.Views
 {
     public class ConfigurationManager : Window
     {
