+++ OpenTabletDriver/Views/LogView.xaml
@@ -3,7 +3,7 @@
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-    x:Class="OpenTabletDriverGUI.Views.LogView">
+    x:Class="OpenTabletDriver.Views.LogView">
     <ListBox Items="{Binding Messages, Mode=TwoWay}">
         <ListBox.ItemTemplate>
             <DataTemplate>
