+++ OpenTabletDriver/Views/MainWindow.xaml
@@ -1,17 +1,17 @@
 <Window xmlns="https://github.com/avaloniaui"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:sys="clr-namesplace:System"
-        xmlns:vm="clr-namespace:OpenTabletDriverGUI.ViewModels;assembly=OpenTabletDriverGUI"
+        xmlns:vm="clr-namespace:OpenTabletDriver.ViewModels;assembly=OpenTabletDriver"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:ns="clr-namespace:OpenTabletDriverGUI.Views;assembly=OpenTabletDriverGUI"
+        xmlns:ns="clr-namespace:OpenTabletDriver.Views;assembly=OpenTabletDriver"
         xmlns:cursor="clr-namespace:TabletDriverLib.Interop.Cursor;assembly=TabletDriverLib"
         mc:Ignorable="d" d:DesignWidth="800" d:DesignHeight="450"
-        x:Class="OpenTabletDriverGUI.Views.MainWindow"
-        Icon="avares://OpenTabletDriverGUI/Assets/otd.ico"
+        x:Class="OpenTabletDriver.Views.MainWindow"
+        Icon="avares://OpenTabletDriver/Assets/otd.ico"
         Width="1280" Height="720"
         MinWidth="640" MinHeight="480"
-        Title="OpenTabletDriverGUI">
+        Title="OpenTabletDriver">
     <Design.DataContext>
         <vm:MainWindowViewModel/>
     </Design.DataContext>
