+++ OpenTabletDriver/Views/StatusBar.xaml
@@ -5,7 +5,7 @@
     mc:Ignorable="d"
     d:DesignWidth="800" d:DesignHeight="450"
     MinHeight="18"
-    x:Class="OpenTabletDriverGUI.Views.StatusBar">
+    x:Class="OpenTabletDriver.Views.StatusBar">
     <UserControl.Styles>
         <Style Selector="Border">
             <Setter Property="BorderBrush" Value="{DynamicResource ThemeBorderMidBrush}"/>
